Исправил и выполнил 2 и 3 часть
