+++ src/ru/gb/jtwo/alesson/online/Animal.java
@@ -2,14 +2,14 @@ package ru.gb.jtwo.alesson.online;
 
 public class Animal {
     int age;
-    String name;
+    public String name;
 
     public Animal(String name) {
         this.name = name;
         age = 10;
     }
 
-    void walk() {
+    public void walk() {
         System.out.println(name + " walks");
     }
 }
